+++ resharper/resharper-unity/src/CSharp/Daemon/UsageChecking/UsageInspectionsSuppressor.cs
@@ -45,7 +45,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.UsageChecking
                     flags = ImplicitUseKindFlags.InstantiatedNoFixedConstructorSignature;
                     return true;
 
-                case ITypeElement typeElement when unityApi.IsSerializableType(typeElement):
+                case ITypeElement typeElement when unityApi.IsSerializableTypeDeclaration(typeElement):
                     // TODO: We should only really mark it as in use if it's actually used somewhere
                     // That is, it should be used as a field in a Unity type, or another serializable type
                     flags = ImplicitUseKindFlags.InstantiatedNoFixedConstructorSignature;